requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.11
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
idna==3.10
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
PyYAML==6.0.2
redis==5.2.1
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
unicorn==2.1.3
uvicorn==0.34.0
